backend/requirements.txt
alembic==1.11.1
anyio==4.13.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.95.1
greenlet==3.4.0
h11==0.16.0
idna==3.11
Mako==1.3.11
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==1.10.26
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.26
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing_extensions==4.15.0
uvicorn==0.22.0
